Too much to say

+++ src/MSL_C/include/string.h
@@ -7,19 +7,19 @@
 extern "C" {
 #endif
 
-char *strcpy(char *, const char *);
-char *strncpy(char *, const char *, size_t);
+char* strcpy(char*, const char*);
+char* strncpy(char*, const char*, unsigned long);
 
-char *strcat(char *, const char *);
-char *strncat(char *, const char *, size_t);
+char* strcat(char*, const char*);
+char* strncat(char*, const char*, unsigned long);
 
-int strcmp(const char *, const char *);
-int strncmp(const char *, const char *, size_t);
+int strcmp(const char*, const char*);
+int strncmp(const char*, const char*, unsigned long);
 
-char *strchr(const char *, int);
-char *strstr(const char *, const char *);
+char* strchr(const char*, int);
+char* strstr(const char*, const char*);
 
-size_t strlen(const char *);
+size_t strlen(const char*);
 
 #ifdef __cplusplus
 }
